@@ -1,10 +1,9 @@
-﻿namespace ConsoleAutoMapper
+﻿namespace ConsoleAutoMapper;
+
+public class User
 {
-    public class User
-    {
-        public string Id { get; set; }
-        public string Name { get; set; }
-        public string Tel { get; set; }
-        public decimal? Balance { get; set; }
-    }
+    public string Id { get; set; }
+    public string Name { get; set; }
+    public string Tel { get; set; }
+    public decimal? Balance { get; set; }
 }
