@@ -1,12 +1,11 @@
 ﻿using Bellight.AutoMapper;
 
-namespace ConsoleAutoMapper
+namespace ConsoleAutoMapper;
+
+[MappedModel(typeof(User))]
+public class UserViewModel
 {
-    [MappedModel(typeof(User))]
-    public class UserViewModel
-    {
-        public string Name { get; set; }
-        public string Tel { get; set; }
-        public string Balance { get; set; }
-    }
+    public string Name { get; set; }
+    public string Tel { get; set; }
+    public string Balance { get; set; }
 }
