@@ -4,55 +4,54 @@ using Microsoft.Extensions.DependencyInjection;
 using Serilog;
 using System;
 
-namespace MessageBusSubscriber
+namespace MessageBusSubscriber;
+
+internal class Program
 {
-    internal class Program
+    private static void Main(string[] args)
     {
-        private static void Main(string[] args)
-        {
-            var topic = "test1";
-            var messageBusType = MessageBusType.PubSub;
+        var topic = "test1";
+        var messageBusType = MessageBusType.PubSub;
 
-            var subscriberName = "sub1";
+        var subscriberName = "sub1";
 
-            var typeText = messageBusType == MessageBusType.Queue ? "queue" : "topic";
-            Console.WriteLine($"Subscribing {typeText} '{topic}'");
+        var typeText = messageBusType == MessageBusType.Queue ? "queue" : "topic";
+        Console.WriteLine($"Subscribing {typeText} '{topic}'");
 
-            // Settings for Azure Message Bus
-            // var policyName = WebUtility.UrlEncode(""); // enter policy name
-            // var key = WebUtility.UrlEncode(""); // enter key
-            // var namespaceUrl = "";
+        // Settings for Azure Message Bus
+        // var policyName = WebUtility.UrlEncode(""); // enter policy name
+        // var key = WebUtility.UrlEncode(""); // enter key
+        // var namespaceUrl = "";
 
-            // var connectionString = $"amqps://{policyName}:{key}@{namespaceUrl}/";
+        // var connectionString = $"amqps://{policyName}:{key}@{namespaceUrl}/";
 
-            var connectionString = "amqp://artemis:simetraehcapa@localhost:5672";
+        var connectionString = "amqp://artemis:simetraehcapa@localhost:5672";
 
-            Console.WriteLine(connectionString);
-            var services = new ServiceCollection();
-            services.AddLogging(loggingBuilder =>
-                loggingBuilder.AddSerilog(dispose: true));
+        Console.WriteLine(connectionString);
+        var services = new ServiceCollection();
+        services.AddLogging(loggingBuilder =>
+            loggingBuilder.AddSerilog(dispose: true));
 
-            services.AddBellightMessageBus()
-                .AddAmqp(options =>
-                {
-                    options.Endpoint = connectionString;
-                    options.IsAzureMessageBus = "false";
-                    options.SubscriberName = subscriberName;
-                });
+        services.AddBellightMessageBus()
+            .AddAmqp(options =>
+            {
+                options.Endpoint = connectionString;
+                options.IsAzureMessageBus = "false";
+                options.SubscriberName = subscriberName;
+            });
 
-            var serviceProvider = services.BuildServiceProvider();
-            serviceProvider.ConfigureCoreLogging();
+        var serviceProvider = services.BuildServiceProvider();
+        serviceProvider.ConfigureCoreLogging();
 
-            var messageBusFactory = serviceProvider.GetService<IMessageBusFactory>();
+        var messageBusFactory = serviceProvider.GetService<IMessageBusFactory>();
 
-            var subscription = messageBusFactory.Subscribe(topic, OnMessageReceived, messageBusType);
-            Console.ReadLine();
-            subscription.Dispose();
-        }
+        var subscription = messageBusFactory.Subscribe(topic, OnMessageReceived, messageBusType);
+        Console.ReadLine();
+        subscription.Dispose();
+    }
 
-        private static void OnMessageReceived(string message)
-        {
-            Console.WriteLine(message);
-        }
+    private static void OnMessageReceived(string message)
+    {
+        Console.WriteLine(message);
     }
 }
