@@ -4,32 +4,31 @@ using Microsoft.AspNetCore.Mvc;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
-namespace MessageBusPublisherWeb.Controllers
+namespace MessageBusPublisherWeb.Controllers;
+
+[Route("api/messages")]
+[ApiController]
+public class MessagesController : ControllerBase
 {
-    [Route("api/messages")]
-    [ApiController]
-    public class MessagesController : ControllerBase
+    private readonly IMessageBusFactory messageBusFactory;
+    private readonly IConfiguration configuration;
+
+    public MessagesController(IMessageBusFactory messageBusFactory, IConfiguration configuration)
     {
-        private readonly IMessageBusFactory messageBusFactory;
-        private readonly IConfiguration configuration;
+        this.messageBusFactory = messageBusFactory;
+        this.configuration = configuration;
+    }
 
-        public MessagesController(IMessageBusFactory messageBusFactory, IConfiguration configuration)
-        {
-            this.messageBusFactory = messageBusFactory;
-            this.configuration = configuration;
-        }
+    // POST api/<MessagesController>
+    [HttpPost]
+    public async Task<IActionResult> Post([FromBody] Message message)
+    {
+        var topic = configuration["ServiceBus:Topic"];
+        var messageBusType = "PubSub".Equals(configuration["ServiceBus:Type"], StringComparison.InvariantCultureIgnoreCase) ?
+            MessageBusType.PubSub : MessageBusType.Queue;
+        var publisher = messageBusFactory.GetPublisher(topic!, messageBusType);
 
-        // POST api/<MessagesController>
-        [HttpPost]
-        public async Task<IActionResult> Post([FromBody] Message message)
-        {
-            var topic = configuration["ServiceBus:Topic"];
-            var messageBusType = "PubSub".Equals(configuration["ServiceBus:Type"], StringComparison.InvariantCultureIgnoreCase) ?
-                MessageBusType.PubSub : MessageBusType.Queue;
-            var publisher = messageBusFactory.GetPublisher(topic!, messageBusType);
-
-            await publisher.SendAsync(message.Content);
-            return Ok();
-        }
+        await publisher.SendAsync(message.Content);
+        return Ok();
     }
 }
