@@ -1,7 +1,6 @@
-﻿namespace MessageBusPublisherWeb.Models
+﻿namespace MessageBusPublisherWeb.Models;
+
+public class Message
 {
-    public class Message
-    {
-        public string Content { get; set; } = string.Empty;
-    }
+    public string Content { get; set; } = string.Empty;
 }
