@@ -1,14 +1,13 @@
 ﻿using Bellight.Core;
 using System;
 
-namespace SimpleIoC
+namespace SimpleIoC;
+
+[KeyedService("C2")]
+public class KeyedServiceC2 : IKeyedServiceC
 {
-    [KeyedService("C2")]
-    public class KeyedServiceC2 : IKeyedServiceC
+    public void DoSomethingInKeyed()
     {
-        public void DoSomethingInKeyed()
-        {
-            Console.WriteLine("IKeyedServiceC - KeyedServiceC2 - DoSomethingInKeyed()");
-        }
+        Console.WriteLine("IKeyedServiceC - KeyedServiceC2 - DoSomethingInKeyed()");
     }
 }
