@@ -8,19 +8,15 @@ public class DefaultDependencyCacheService(
     ISerializer serializer,
     IServiceProvider serviceProvider) : IDependencyCacheService
 {
-    private readonly BellightCoreOptions _options = options;
-    private readonly ISerializer _serializer = serializer;
-    private readonly IServiceProvider _serviceProvider = serviceProvider;
-
     public bool Load()
     {
-        if (_options.DependencyCacheOptions?.Enabled != true)
+        if (options.DependencyCacheOptions?.Enabled != true)
         {
             return false;
         }
 
-        var filePath = string.IsNullOrEmpty(_options.DependencyCacheOptions.FileLocation) ? _options.DependencyCacheOptions.FileName :
-            Path.Combine(_options.DependencyCacheOptions.FileLocation, _options.DependencyCacheOptions.FileName);
+        var filePath = string.IsNullOrEmpty(options.DependencyCacheOptions.FileLocation) ? options.DependencyCacheOptions.FileName :
+            Path.Combine(options.DependencyCacheOptions.FileLocation, options.DependencyCacheOptions.FileName);
 
         var content = LoadCache(filePath);
         if (string.IsNullOrEmpty(content))
@@ -28,9 +24,9 @@ public class DefaultDependencyCacheService(
             return false;
         }
 
-        var model = _serializer.TryDeserializeObject<DependencyCacheModel>(content);
+        var model = serializer.TryDeserializeObject<DependencyCacheModel>(content);
 
-        if (model == null || !VerifyAssembles(model.Assemblies!))
+        if (!VerifyAssembles(model.Assemblies!))
         {
             return false;
         }
@@ -45,7 +41,7 @@ public class DefaultDependencyCacheService(
             foreach (var item in model.TypeHandlers)
             {
                 var type = Type.GetType(item.Name!);
-                var handler = _serviceProvider.GetService(type!) as ITypeHandler;
+                var handler = serviceProvider.GetService(type!) as ITypeHandler;
                 handler?.LoadCache(item.Sections!);
             }
         });
@@ -55,24 +51,24 @@ public class DefaultDependencyCacheService(
 
     public void Save(DependencyCacheModel item)
     {
-        if (item == null || _options.DependencyCacheOptions?.Enabled != true)
+        if (options.DependencyCacheOptions?.Enabled != true)
         {
             return;
         }
 
-        if (_options.DependencyCacheOptions.PrettyPrint)
+        if (options.DependencyCacheOptions.PrettyPrint)
         {
-            _serializer.Settings.WriteIndented = true;
+            serializer.Settings.WriteIndented = true;
         }
 
-        var serializedContent = _serializer.TrySerializeObject(item);
+        var serializedContent = serializer.TrySerializeObject(item);
         if (string.IsNullOrEmpty(serializedContent))
         {
             return;
         }
 
-        var filePath = string.IsNullOrEmpty(_options.DependencyCacheOptions.FileLocation) ? _options.DependencyCacheOptions.FileName :
-            Path.Combine(_options.DependencyCacheOptions.FileLocation, _options.DependencyCacheOptions.FileName);
+        var filePath = string.IsNullOrEmpty(options.DependencyCacheOptions.FileLocation) ? options.DependencyCacheOptions.FileName :
+            Path.Combine(options.DependencyCacheOptions.FileLocation, options.DependencyCacheOptions.FileName);
 
         try
         {
@@ -80,7 +76,7 @@ public class DefaultDependencyCacheService(
         }
         catch
         {
-            CoreLogging.Logger?.LogWarning("Cannot write cache file: {filePath}. As a consequence, next startup may suffer a performance issue.", filePath);
+            CoreLogging.Logger?.LogWarning("Cannot write cache file: {FilePath}. As a consequence, next startup may suffer a performance issue.", filePath);
         }
     }
 
@@ -97,7 +93,7 @@ public class DefaultDependencyCacheService(
         }
         catch
         {
-            CoreLogging.Logger?.LogWarning("Cannot open cache file: {fileName}", fileName);
+            CoreLogging.Logger?.LogWarning("Cannot open cache file: {FileName}", fileName);
             return string.Empty;
         }
     }
