@@ -6,19 +6,17 @@ namespace Bellight.MongoDb.Transactions;
 
 public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
 {
-    private readonly IMongoDatabase _database = database;
-
     public IAsyncCursor<TResult> Aggregate<TResult>(
         PipelineDefinition<NoPipelineInput, TResult> pipeline,
         AggregateOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
             return Aggregate(session, pipeline, options, cancellationToken);
         }
 
-        return _database.Aggregate(pipeline, options, cancellationToken);
+        return database.Aggregate(pipeline, options, cancellationToken);
     }
 
     public IAsyncCursor<TResult> Aggregate<TResult>(
@@ -27,7 +25,7 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         AggregateOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        return _database.Aggregate(session, pipeline, options, cancellationToken);
+        return database.Aggregate(session, pipeline, options, cancellationToken);
     }
 
     public Task<IAsyncCursor<TResult>> AggregateAsync<TResult>(
@@ -35,12 +33,12 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         AggregateOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            return _database.AggregateAsync(session, pipeline, options, cancellationToken);
+            return database.AggregateAsync(session, pipeline, options, cancellationToken);
         }
 
-        return _database.AggregateAsync(pipeline, options, cancellationToken);
+        return database.AggregateAsync(pipeline, options, cancellationToken);
     }
 
     public Task<IAsyncCursor<TResult>> AggregateAsync<TResult>(
@@ -49,7 +47,7 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         AggregateOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        return _database.AggregateAsync(session, pipeline, options, cancellationToken);
+        return database.AggregateAsync(session, pipeline, options, cancellationToken);
     }
 
     public void AggregateToCollection<TResult>(
@@ -57,13 +55,13 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         AggregateOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            _database.AggregateToCollection(session, pipeline, options, cancellationToken);
+            database.AggregateToCollection(session, pipeline, options, cancellationToken);
             return;
         }
 
-        _database.AggregateToCollection(pipeline, options, cancellationToken);
+        database.AggregateToCollection(pipeline, options, cancellationToken);
     }
 
     public void AggregateToCollection<TResult>(
@@ -72,7 +70,7 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         AggregateOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        _database.AggregateToCollection(session, pipeline, options, cancellationToken);
+        database.AggregateToCollection(session, pipeline, options, cancellationToken);
     }
 
     public Task AggregateToCollectionAsync<TResult>(
@@ -80,13 +78,12 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         AggregateOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            return _database
-                .AggregateToCollectionAsync(session, pipeline, options, cancellationToken);
+            return database.AggregateToCollectionAsync(session, pipeline, options, cancellationToken);
         }
 
-        return _database.AggregateToCollectionAsync(pipeline, options, cancellationToken);
+        return database.AggregateToCollectionAsync(pipeline, options, cancellationToken);
     }
 
     public Task AggregateToCollectionAsync<TResult>(
@@ -95,8 +92,7 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         AggregateOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        return _database
-            .AggregateToCollectionAsync(session, pipeline, options, cancellationToken);
+        return database.AggregateToCollectionAsync(session, pipeline, options, cancellationToken);
     }
 
     public void CreateCollection(
@@ -104,13 +100,13 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         CreateCollectionOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            _database.CreateCollection(session, name, options, cancellationToken);
+            database.CreateCollection(session, name, options, cancellationToken);
             return;
         }
 
-        _database.CreateCollection(name, options, cancellationToken);
+        database.CreateCollection(name, options, cancellationToken);
     }
 
     public void CreateCollection(
@@ -119,7 +115,7 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         CreateCollectionOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        _database.CreateCollection(session, name, options, cancellationToken);
+        database.CreateCollection(session, name, options, cancellationToken);
     }
 
     public Task CreateCollectionAsync(
@@ -127,12 +123,12 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         CreateCollectionOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            return _database.CreateCollectionAsync(session, name, options, cancellationToken);
+            return database.CreateCollectionAsync(session, name, options, cancellationToken);
         }
 
-        return _database.CreateCollectionAsync(name, options, cancellationToken);
+        return database.CreateCollectionAsync(name, options, cancellationToken);
     }
 
     public Task CreateCollectionAsync(
@@ -141,7 +137,7 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         CreateCollectionOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        return _database.CreateCollectionAsync(session, name, options, cancellationToken);
+        return database.CreateCollectionAsync(session, name, options, cancellationToken);
     }
 
     public void CreateView<TDocument, TResult>(
@@ -151,14 +147,13 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         CreateViewOptions<TDocument>? options = null,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            _database
-                .CreateView(session, viewName, viewOn, pipeline, options, cancellationToken);
+            database.CreateView(session, viewName, viewOn, pipeline, options, cancellationToken);
             return;
         }
 
-        _database.CreateView(viewName, viewOn, pipeline, options, cancellationToken);
+        database.CreateView(viewName, viewOn, pipeline, options, cancellationToken);
     }
 
     public void CreateView<TDocument, TResult>(
@@ -169,7 +164,7 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         CreateViewOptions<TDocument>? options = null,
         CancellationToken cancellationToken = default)
     {
-        _database.CreateView(session, viewName, viewOn, pipeline, options, cancellationToken);
+        database.CreateView(session, viewName, viewOn, pipeline, options, cancellationToken);
     }
 
     public Task CreateViewAsync<TDocument, TResult>(
@@ -179,9 +174,9 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         CreateViewOptions<TDocument>? options = null,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            return _database.CreateViewAsync(
+            return database.CreateViewAsync(
                 session,
                 viewName,
                 viewOn,
@@ -190,8 +185,7 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
                 cancellationToken);
         }
 
-        return _database
-            .CreateViewAsync(viewName, viewOn, pipeline, options, cancellationToken);
+        return database.CreateViewAsync(viewName, viewOn, pipeline, options, cancellationToken);
     }
 
     public Task CreateViewAsync<TDocument, TResult>(
@@ -202,19 +196,18 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         CreateViewOptions<TDocument>? options = null,
         CancellationToken cancellationToken = default)
     {
-        return _database
-            .CreateViewAsync(session, viewName, viewOn, pipeline, options, cancellationToken);
+        return database.CreateViewAsync(session, viewName, viewOn, pipeline, options, cancellationToken);
     }
 
     public void DropCollection(string name, CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            _database.DropCollection(session, name, cancellationToken);
+            database.DropCollection(session, name, cancellationToken);
             return;
         }
 
-        _database.DropCollection(name, cancellationToken);
+        database.DropCollection(name, cancellationToken);
     }
 
     public void DropCollection(
@@ -222,29 +215,35 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         string name,
         CancellationToken cancellationToken = default)
     {
-        _database.DropCollection(session, name, cancellationToken);
+        database.DropCollection(session, name, cancellationToken);
     }
 
     public void DropCollection(string name, DropCollectionOptions options, CancellationToken cancellationToken = default)
     {
-        _database.DropCollection(name, options, cancellationToken);
+        if (TryGetSession(out var session))
+        {
+            database.DropCollection(session, name, options, cancellationToken);
+            return;
+        }
+
+        database.DropCollection(name, options, cancellationToken);
     }
 
     public void DropCollection(IClientSessionHandle session, string name, DropCollectionOptions options, CancellationToken cancellationToken = default)
     {
-        _database.DropCollection(session, name, options, cancellationToken);
+        database.DropCollection(session, name, options, cancellationToken);
     }
 
     public Task DropCollectionAsync(
         string name,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            return _database.DropCollectionAsync(session, name, cancellationToken);
+            return database.DropCollectionAsync(session, name, cancellationToken);
         }
 
-        return _database.DropCollectionAsync(name, cancellationToken);
+        return database.DropCollectionAsync(name, cancellationToken);
     }
 
     public Task DropCollectionAsync(
@@ -252,36 +251,36 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         string name,
         CancellationToken cancellationToken = default)
     {
-        return _database.DropCollectionAsync(session, name, cancellationToken);
+        return database.DropCollectionAsync(session, name, cancellationToken);
     }
 
     public Task DropCollectionAsync(string name, DropCollectionOptions options, CancellationToken cancellationToken = default)
     {
-        return _database.DropCollectionAsync(name, options, cancellationToken);
+        return database.DropCollectionAsync(name, options, cancellationToken);
     }
 
     public Task DropCollectionAsync(IClientSessionHandle session, string name, DropCollectionOptions options, CancellationToken cancellationToken = default)
     {
-        return _database.DropCollectionAsync(session, name, options, cancellationToken);
+        return database.DropCollectionAsync(session, name, options, cancellationToken);
     }
 
     public IMongoCollection<TDocument> GetCollection<TDocument>(
         string name,
         MongoCollectionSettings? settings = null)
     {
-        return _database.GetCollection<TDocument>(name, settings).AsTransactionCollection();
+        return database.GetCollection<TDocument>(name, settings).AsTransactionCollection();
     }
 
     public IAsyncCursor<string> ListCollectionNames(
         ListCollectionNamesOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            return _database.ListCollectionNames(session, options, cancellationToken);
+            return database.ListCollectionNames(session, options, cancellationToken);
         }
 
-        return _database.ListCollectionNames(options, cancellationToken);
+        return database.ListCollectionNames(options, cancellationToken);
     }
 
     public IAsyncCursor<string> ListCollectionNames(
@@ -289,19 +288,19 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         ListCollectionNamesOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        return _database.ListCollectionNames(session, options, cancellationToken);
+        return database.ListCollectionNames(session, options, cancellationToken);
     }
 
     public Task<IAsyncCursor<string>> ListCollectionNamesAsync(
         ListCollectionNamesOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            return _database.ListCollectionNamesAsync(session, options, cancellationToken);
+            return database.ListCollectionNamesAsync(session, options, cancellationToken);
         }
 
-        return _database.ListCollectionNamesAsync(options, cancellationToken);
+        return database.ListCollectionNamesAsync(options, cancellationToken);
     }
 
     public Task<IAsyncCursor<string>> ListCollectionNamesAsync(
@@ -309,19 +308,19 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         ListCollectionNamesOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        return _database.ListCollectionNamesAsync(session, options, cancellationToken);
+        return database.ListCollectionNamesAsync(session, options, cancellationToken);
     }
 
     public IAsyncCursor<BsonDocument> ListCollections(
         ListCollectionsOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            return _database.ListCollections(session, options, cancellationToken);
+            return database.ListCollections(session, options, cancellationToken);
         }
 
-        return _database.ListCollections(options, cancellationToken);
+        return database.ListCollections(options, cancellationToken);
     }
 
     public IAsyncCursor<BsonDocument> ListCollections(
@@ -329,19 +328,19 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         ListCollectionsOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        return _database.ListCollections(session, options, cancellationToken);
+        return database.ListCollections(session, options, cancellationToken);
     }
 
     public Task<IAsyncCursor<BsonDocument>> ListCollectionsAsync(
         ListCollectionsOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            return _database.ListCollectionsAsync(session, options, cancellationToken);
+            return database.ListCollectionsAsync(session, options, cancellationToken);
         }
 
-        return _database.ListCollectionsAsync(options, cancellationToken);
+        return database.ListCollectionsAsync(options, cancellationToken);
     }
 
     public Task<IAsyncCursor<BsonDocument>> ListCollectionsAsync(
@@ -349,7 +348,7 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         ListCollectionsOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        return _database.ListCollectionsAsync(session, options, cancellationToken);
+        return database.ListCollectionsAsync(session, options, cancellationToken);
     }
 
     public void RenameCollection(
@@ -358,13 +357,13 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         RenameCollectionOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            _database.RenameCollection(session, oldName, newName, options, cancellationToken);
+            database.RenameCollection(session, oldName, newName, options, cancellationToken);
             return;
         }
 
-        _database.RenameCollection(oldName, newName, options, cancellationToken);
+        database.RenameCollection(oldName, newName, options, cancellationToken);
     }
 
     public void RenameCollection(
@@ -374,7 +373,7 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         RenameCollectionOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        _database.RenameCollection(session, oldName, newName, options, cancellationToken);
+        database.RenameCollection(session, oldName, newName, options, cancellationToken);
     }
 
     public Task RenameCollectionAsync(
@@ -383,13 +382,12 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         RenameCollectionOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            return _database
-                .RenameCollectionAsync(session, oldName, newName, options, cancellationToken);
+            return database.RenameCollectionAsync(session, oldName, newName, options, cancellationToken);
         }
 
-        return _database.RenameCollectionAsync(oldName, newName, options, cancellationToken);
+        return database.RenameCollectionAsync(oldName, newName, options, cancellationToken);
     }
 
     public Task RenameCollectionAsync(
@@ -399,8 +397,7 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         RenameCollectionOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        return _database
-            .RenameCollectionAsync(session, oldName, newName, options, cancellationToken);
+        return database.RenameCollectionAsync(session, oldName, newName, options, cancellationToken);
     }
 
     public TResult RunCommand<TResult>(
@@ -408,12 +405,12 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         ReadPreference? readPreference = null,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            return _database.RunCommand(session, command, readPreference, cancellationToken);
+            return database.RunCommand(session, command, readPreference, cancellationToken);
         }
 
-        return _database.RunCommand(command, readPreference, cancellationToken);
+        return database.RunCommand(command, readPreference, cancellationToken);
     }
 
     public TResult RunCommand<TResult>(
@@ -422,7 +419,7 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         ReadPreference? readPreference = null,
         CancellationToken cancellationToken = default)
     {
-        return _database.RunCommand(session, command, readPreference, cancellationToken);
+        return database.RunCommand(session, command, readPreference, cancellationToken);
     }
 
     public Task<TResult> RunCommandAsync<TResult>(
@@ -430,7 +427,7 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         ReadPreference? readPreference = null,
         CancellationToken cancellationToken = default)
     {
-        return _database.RunCommandAsync(command, readPreference, cancellationToken);
+        return database.RunCommandAsync(command, readPreference, cancellationToken);
     }
 
     public Task<TResult> RunCommandAsync<TResult>(
@@ -439,7 +436,7 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         ReadPreference? readPreference = null,
         CancellationToken cancellationToken = default)
     {
-        return _database.RunCommandAsync(session, command, readPreference, cancellationToken);
+        return database.RunCommandAsync(session, command, readPreference, cancellationToken);
     }
 
     public IChangeStreamCursor<TResult> Watch<TResult>(
@@ -447,12 +444,12 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         ChangeStreamOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            return _database.Watch(session, pipeline, options, cancellationToken);
+            return database.Watch(session, pipeline, options, cancellationToken);
         }
 
-        return _database.Watch(pipeline, options, cancellationToken);
+        return database.Watch(pipeline, options, cancellationToken);
     }
 
     public IChangeStreamCursor<TResult> Watch<TResult>(
@@ -461,7 +458,7 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         ChangeStreamOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        return _database.Watch(session, pipeline, options, cancellationToken);
+        return database.Watch(session, pipeline, options, cancellationToken);
     }
 
     public Task<IChangeStreamCursor<TResult>> WatchAsync<TResult>(
@@ -469,12 +466,12 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         ChangeStreamOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        if (TryGetSession(out IClientSessionHandle? session))
+        if (TryGetSession(out var session))
         {
-            return _database.WatchAsync(session, pipeline, options, cancellationToken);
+            return database.WatchAsync(session, pipeline, options, cancellationToken);
         }
 
-        return _database.WatchAsync(pipeline, options, cancellationToken);
+        return database.WatchAsync(pipeline, options, cancellationToken);
     }
 
     public Task<IChangeStreamCursor<TResult>> WatchAsync<TResult>(
@@ -483,31 +480,31 @@ public class MongoTransactionDatabase(IMongoDatabase database) : IMongoDatabase
         ChangeStreamOptions? options = null,
         CancellationToken cancellationToken = default)
     {
-        return _database.WatchAsync(session, pipeline, options, cancellationToken);
+        return database.WatchAsync(session, pipeline, options, cancellationToken);
     }
 
     public IMongoDatabase WithReadConcern(ReadConcern readConcern)
     {
-        return _database.WithReadConcern(readConcern).AsTransactionDatabase();
+        return database.WithReadConcern(readConcern).AsTransactionDatabase();
     }
 
     public IMongoDatabase WithReadPreference(ReadPreference readPreference)
     {
-        return _database.WithReadPreference(readPreference).AsTransactionDatabase();
+        return database.WithReadPreference(readPreference).AsTransactionDatabase();
     }
 
     public IMongoDatabase WithWriteConcern(WriteConcern writeConcern)
     {
-        return _database.WithWriteConcern(writeConcern).AsTransactionDatabase();
+        return database.WithWriteConcern(writeConcern).AsTransactionDatabase();
     }
 
-    public IMongoClient Client => _database.Client;
+    public IMongoClient Client => database.Client;
 
-    public DatabaseNamespace DatabaseNamespace => _database.DatabaseNamespace;
+    public DatabaseNamespace DatabaseNamespace => database.DatabaseNamespace;
 
-    public MongoDatabaseSettings Settings => _database.Settings;
+    public MongoDatabaseSettings Settings => database.Settings;
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     private bool TryGetSession(out IClientSessionHandle sessionHandle) =>
-        TransactionStore.TryGetSession(_database.Client, out sessionHandle);
+        TransactionStore.TryGetSession(database.Client, out sessionHandle);
 }
